Fixed some bugs
1. The SAX parser sometimes breaks up paths, if they are too long.
Fixed it by cumulating data until </filename> is encountered.
2. When executing Open Project, a Save File dialog came up. Replaced it
by a Open File dialog.

+++ src/ui/panels/version/VersionPanel.java
@@ -210,7 +210,7 @@ public class VersionPanel extends Panel
 		int subSampleMode		= subSampleModeCB.getSelectedIndex();
 		int numberOfSamplesMode	= numberOfSamplesModeCB.getSelectedIndex();
 		int subSampleSize		= Integer.parseInt(subSampleSizeText.getText());
-		double numberOfSamples	= Integer.parseInt(numberOfSamplesText.getText());
+		double numberOfSamples	= Double.parseDouble(numberOfSamplesText.getText());
 		
 		try {
 			versionCell.setValue(weighting,
